Refactorizacion TemplateService
